+++ src/evaluate.ts
@@ -53,7 +53,7 @@ async function evaluateCode(code: string, options, selection?: vscode.Selection)
     const ns = options.ns;
 
     if (code.length > 0) {
-        let err: string[] = [], out: string[] = [];
+        let err: string[] = [];
 
         if (outputWindow.getNs() !== ns) {
             await session.eval("(in-ns '" + ns + ")", session.client.ns).value;
@@ -64,13 +64,12 @@ async function evaluateCode(code: string, options, selection?: vscode.Selection)
             line: line + 1,
             column: column + 1,
             stdout: (m) => {
-                out.push(m);
                 outputWindow.append(normalizeNewLines(m));
             },
             stderr: m => err.push(m),
             pprintOptions: pprintOptions
         });
-        
+
         try {
             let value = await context.value;
             value = util.stripAnsi(context.pprintOut || value);
@@ -87,9 +86,10 @@ async function evaluateCode(code: string, options, selection?: vscode.Selection)
                     } else if (options.comment) {
                         addAsComment(c, value, selection, editor, selection);
                     } else {
-                        const currentCursorPos = editor.selection.active;
-                        annotations.decorateSelection(value, selection, editor, currentCursorPos, resultLocation, annotations.AnnotationStatus.SUCCESS);
-                        annotations.decorateResults(value, false, selection, editor);
+                        if (!outputWindow.isResultsDoc(editor.document)) {
+                            annotations.decorateSelection(value, selection, editor, editor.selection.active, resultLocation, annotations.AnnotationStatus.SUCCESS);
+                            annotations.decorateResults(value, false, selection, editor);
+                        }
                     }
                 }
             });
@@ -97,28 +97,31 @@ async function evaluateCode(code: string, options, selection?: vscode.Selection)
             if (err.length > 0) {
                 outputWindow.append(`; ${normalizeNewLinesAndJoin(err, true)}`);
                 if (context.stacktrace) {
-                    outputWindow.printStacktrace(context.stacktrace);
+                    outputWindow.saveStacktrace(context.stacktrace);
+                    outputWindow.printLastStacktrace();
                 }
             }
         } catch (e) {
             const outputWindowError = err.length ? `; ${normalizeNewLinesAndJoin(err, true)}` : formatAsLineComments(e);
-            outputWindow.append(outputWindowError, (resultLocation) => {
+            outputWindow.append(outputWindowError, async (resultLocation) => {
                 if (selection) {
                     const editor = vscode.window.activeTextEditor;
                     const editorError = util.stripAnsi(err.length ? err.join("\n") : e);
                     const currentCursorPos = editor.selection.active;
-                    annotations.decorateSelection(editorError, selection, editor, currentCursorPos, resultLocation, annotations.AnnotationStatus.ERROR);
-                    annotations.decorateResults(editorError, true, selection, editor);
+                    if (!outputWindow.isResultsDoc(editor.document)) {
+                        annotations.decorateSelection(editorError, selection, editor, currentCursorPos, resultLocation, annotations.AnnotationStatus.ERROR);
+                        annotations.decorateResults(editorError, true, selection, editor);
+                    }
                     if (options.asComment) {
                         addAsComment(selection.start.character, editorError, selection, editor, selection);
                     }
                 }
-                if (context.stacktrace && context.stacktrace.stacktrace) {
-                    outputWindow.printStacktrace(context.stacktrace.stacktrace);
-                }
             });
+            if (context.stacktrace && context.stacktrace.stacktrace) {
+                outputWindow.saveStacktrace(context.stacktrace.stacktrace);
+                outputWindow.printLastStacktrace();
+            }
         }
-
         outputWindow.setSession(session, context.ns || ns);
         namespace.updateREPLSessionType();
     }
@@ -160,6 +163,7 @@ async function evaluateSelection(document: {}, options) {
             }
             annotations.decorateSelection("", codeSelection, editor, undefined, undefined, annotations.AnnotationStatus.PENDING);
             await evaluateCode(code, { ...options, ns, line, column, filePath, session }, codeSelection);
+            outputWindow.appendPrompt();
         }
     } else {
         vscode.window.showErrorMessage("Not connected to a REPL");
@@ -204,14 +208,13 @@ function evaluateCurrentForm(document = {}, options = {}) {
         .catch(printWarningForError);
 }
 
-async function loadFile(document, callback: () => { }, pprintOptions: PrettyPrintingOptions) {
+async function loadFile(document, pprintOptions: PrettyPrintingOptions) {
     const current = state.deref();
     const doc = util.getDocument(document);
     const fileName = util.getFileName(doc);
     const fileType = util.getFileType(doc);
     const ns = namespace.getNamespace(doc);
     const session = namespace.getSession(util.getFileType(doc));
-    const chan = state.outputChannel();
     const shortFileName = path.basename(fileName);
     const dirName = path.dirname(fileName);
 
@@ -221,34 +224,29 @@ async function loadFile(document, callback: () => { }, pprintOptions: PrettyPrin
 
         await session.eval("(in-ns '" + ns + ")", session.client.ns).value;
 
-        let res = session.loadFile(doc.getText(), {
+        const res = session.loadFile(doc.getText(), {
             fileName: fileName,
             filePath: doc.fileName,
             stdout: m => outputWindow.append(normalizeNewLines(m.indexOf(dirName) < 0 ? m.replace(shortFileName, fileName) : m)),
             stderr: m => outputWindow.append('; ' + normalizeNewLines(m.indexOf(dirName) < 0 ? m.replace(shortFileName, fileName) : m, true)),
             pprintOptions: pprintOptions
-        })
-        await res.value.then((value) => {
+        });
+        try {
+            const value = await res.value;
             if (value) {
                 outputWindow.append(value);
             } else {
                 outputWindow.append("; No results from file evaluation.");
             }
-        }).catch(async (e) => {
+        } catch (e) {
             outputWindow.append(`; Evaluation of file ${fileName} failed: ${e}`);
             if (res.stacktrace) {
-                outputWindow.printStacktrace(res.stacktrace);
+                outputWindow.saveStacktrace(res.stacktrace);
+                outputWindow.printLastStacktrace();
             }
-        });
+        }
         outputWindow.setSession(session, res.ns || ns);
         namespace.updateREPLSessionType();
-    }
-    if (callback) {
-        try {
-            callback();
-        } catch (e) {
-            chan.appendLine(`After evaluation callback for file ${fileName} failed: ${e}`);
-        };
     }
 }
 
@@ -339,26 +337,26 @@ async function evaluateInOutputWindow(code: string, sessionType: string, ns: str
         });
     }
     catch (e) {
-        outputWindow.append("; Evaluation failed.")
+        outputWindow.append("; Evaluation failed.");
     }
 }
 
 export type customREPLCommandSnippet = { name: string, snippet: string, repl: string, ns?: string };
 
-function evaluateCustomCommandSnippetCommand() {
-    let pickCounter = 1,
-        configErrors: { "name": string, "keys": string[] }[] = [];
-    const snippets = state.config().customREPLCommandSnippets as customREPLCommandSnippet[],
-        snippetPicks = _.map(snippets, (c: customREPLCommandSnippet) => {
-            const undefs = ["name", "snippet", "repl"].filter(k => {
-                return !c[k];
-            })
-            if (undefs.length > 0) {
-                configErrors.push({ "name": c.name, "keys": undefs });
-            }
-            return `${pickCounter++}: ${c.name} (${c.repl})`;
-        }),
-        snippetsDict = {};
+async function evaluateCustomCommandSnippetCommand(): Promise<void> {
+    let pickCounter = 1;
+    let configErrors: { "name": string, "keys": string[] }[] = [];
+    const snippets = state.config().customREPLCommandSnippets as customREPLCommandSnippet[];
+    const snippetPicks = _.map(snippets, (c: customREPLCommandSnippet) => {
+        const undefs = ["name", "snippet", "repl"].filter(k => {
+            return !c[k];
+        });
+        if (undefs.length > 0) {
+            configErrors.push({ "name": c.name, "keys": undefs });
+        }
+        return `${pickCounter++}: ${c.name} (${c.repl})`;
+    });
+    const snippetsDict = {};
     pickCounter = 1;
 
     if (configErrors.length > 0) {
@@ -370,20 +368,23 @@ function evaluateCustomCommandSnippetCommand() {
     });
 
     if (snippets && snippets.length > 0) {
-        util.quickPickSingle({
-            values: snippetPicks,
-            placeHolder: "Choose a command to run at the REPL",
-            saveAs: "runCustomREPLCommand"
-        }).then(async (pick) => {
+        try {
+            const pick = await util.quickPickSingle({
+                values: snippetPicks,
+                placeHolder: "Choose a command to run at the REPL",
+                saveAs: "runCustomREPLCommand"
+            });
             if (pick && snippetsDict[pick] && snippetsDict[pick].snippet) {
-                const command = snippetsDict[pick].snippet,
-                    editor = vscode.window.activeTextEditor,
-                    editorNS = editor && editor.document && editor.document.languageId === 'clojure' ? namespace.getNamespace(editor.document) : undefined,
-                    ns = snippetsDict[pick].ns ? snippetsDict[pick].ns : editorNS,
-                    repl = snippetsDict[pick].repl ? snippetsDict[pick].repl : "clj";
-                evaluateInOutputWindow(command, repl ? repl : "clj", ns);
+                const command = snippetsDict[pick].snippet;
+                const editor = vscode.window.activeTextEditor;
+                const editorNS = editor && editor.document && editor.document.languageId === 'clojure' ? namespace.getNamespace(editor.document) : undefined;
+                const ns = snippetsDict[pick].ns ? snippetsDict[pick].ns : editorNS;
+                const repl = snippetsDict[pick].repl ? snippetsDict[pick].repl : "clj";
+                await evaluateInOutputWindow(command, repl ? repl : "clj", ns);
             }
-        }).catch(() => { });
+        } catch (e) {
+            console.error(e);
+        }
     } else {
         vscode.window.showInformationMessage("No snippets configured. Configure snippets in `calva.customREPLCommandSnippets`.", ...["OK"]);
     }